Move DebugControl and SingleStep function into debug_ctl.go
